+++ Leomard/Views/Popups/PostPopup.swift
@@ -56,63 +56,7 @@ struct PostPopup: View {
                     .padding(.top, 10)
                     .padding(.bottom, 0)
                     List {
-                        PostUIView(postView: postView, shortBody: false, postService: self.postService, myself: $myself, contentView: contentView)
-                            .frame(
-                                minHeight: 0,
-                                alignment: .top
-                            )
-                        Spacer()
-                        VStack {
-                            Text("Comment")
-                                .frame(
-                                    maxWidth: .infinity,
-                                    alignment: .leading
-                                )
-                                .fontWeight(.semibold)
-                            MarkdownEditor(bodyText: $commentText, contentView: self.contentView)
-                                .frame(
-                                    maxWidth: .infinity,
-                                    minHeight: 3 * NSFont.preferredFont(forTextStyle: .body).xHeight,
-                                    maxHeight: .infinity
-                                )
-                            Button("Send", action: createComment)
-                                .buttonStyle(.borderedProminent)
-                                .frame(
-                                    maxWidth: .infinity,
-                                    alignment: .leading
-                                )
-                                .disabled(isTextFieldEmpty())
-                            if isSendingComment {
-                                ProgressView()
-                                    .progressViewStyle(.circular)
-                            }
-                        }
-                        .frame(maxHeight: .infinity)
-                        Spacer()
-                        Picker("Sort By", selection: $sortType) {
-                            ForEach(CommentSortType.allCases, id: \.self) { method in
-                                Text(String(describing: method))
-                            }
-                        }
-                        .onChange(of: sortType) { value in
-                            page = 1
-                            self.comments = []
-                            loadComments()
-                        }
-                        .frame(maxWidth: 150)
-                        Spacer()
-                        ForEach(comments, id: \.self) { commentView in
-                            CommentUIView(commentView: commentView, indentLevel: 0, commentService: commentService, myself: $myself, post: postView.post, contentView: contentView)
-                                .onAppear {
-                                    if commentView == self.comments.last {
-                                        loadComments()
-                                    }
-                                }
-                            if !(commentView == self.comments.last && lastPage) {
-                                Divider()
-                            }
-                            Spacer()
-                        }
+                        PostOpenedView(postView: postView, contentView: contentView, commentService: commentService, postService: postService, myself: $myself)
                     }
                     .frame(
                         maxWidth: .infinity,
@@ -134,31 +78,7 @@ struct PostPopup: View {
             }
             .cornerRadius(8)
             .padding(.top, 20)
-            .padding(.bottom, 20)
-            .task {
-                self.loadComments()
-                
-                if UserPreferences.getInstance.markPostAsReadOnOpen && !postView.read {
-                    postService.markAsRead(post: postView.post, read: true) { result in
-                        switch result {
-                        case .success(let postResponse):
-                            self.postView = postResponse.postView
-                        case .failure(let error):
-                            print(error)
-                            // TODO: Show error
-                        }
-                    }
-                }
-            }
-            .alert("Error", isPresented: $showingAlert, actions: {
-                Button("Retry", role: .destructive) {
-                    self.createComment()
-                    showingAlert = false
-                }
-                Button("Cancel", role: .cancel) {}
-            }, message: {
-                Text("Failed to send a comment. Try again.")
-             })
+            .padding(.bottom, 20)            
         }
         .frame(
             minWidth: 0,
@@ -171,45 +91,5 @@ struct PostPopup: View {
     func close() {
         contentView.closePost()
         page = 1
-    }
-    
-    func loadComments() {        
-        self.commentService.getAllComments(post: postView.post, page: page, sortType: sortType) { result in
-            switch result {
-            case .success(let getCommentView) :
-                self.comments += getCommentView.comments.filter { !self.comments.contains($0) }
-                page += 1
-            case .failure(let error):
-                print(error)
-            }
-        }
-    }
-    
-    func createComment() {
-        let comment = commentText
-        
-        if isSendingComment {
-            return
-        }
-        
-        isSendingComment = true
-        
-        commentService.createComment(content: comment, post: postView.post) { result in
-            switch result {
-            case .success(let commentResponse):
-                DispatchQueue.main.sync {
-                    comments.insert(commentResponse.commentView, at: 0)
-                    commentText = ""
-                    isSendingComment = false
-                }
-            case .failure(let error):
-                print(error)
-                showingAlert = true
-            }
-        }
-    }
-    
-    func isTextFieldEmpty() -> Bool {
-        return commentText.count == 0
-    }
+    }    
 }